Fix panic in dual stack environment
