+++ pkg/proxy/proxy.go
@@ -153,6 +153,7 @@ func (p *proxy) addEndpointRules(epRule *nftables.EPRule, tableFamily utilnftabl
 // updateServiceChain programs rules for a specific ServicePortName, it is called for every endpoint add/delete
 // event.
 func (p *proxy) updateServiceChain(svcPortName ServicePortName, tableFamily utilnftables.TableFamily) error {
+	klog.V(6).Infof("updating service chain for service %s address family %v", svcPortName.String(), tableFamily)
 	svc, ok := p.serviceMap[svcPortName]
 	if !ok {
 		return nil
@@ -177,6 +178,10 @@ func (p *proxy) updateServiceChain(svcPortName ServicePortName, tableFamily util
 	// Programming rules for existing endpoints
 	epsChains := p.getServicePortEndpointChains(svcPortName, tableFamily)
 	svcRules := entry.svcnft.Chains[tableFamily].Chain[nftables.K8sSvcPrefix+entry.svcnft.ServiceID]
+	if svcRules == nil {
+		klog.Errorf("updating service chain for service %s address family %v failed as Rules array is nil, it is a bug, please file an issue.", svcPortName.String(), tableFamily)
+		return nil
+	}
 	// Check if the service still has any backends
 	if len(epsChains) != 0 {
 		rules, err := nftables.ProgramServiceEndpoints(p.nfti, tableFamily, entry.svcnft.ServiceID, epsChains, svcRules.RuleID, entry.svcnft.WithAffinity, svcPortName.String())
@@ -196,63 +201,5 @@ func (p *proxy) updateServiceChain(svcPortName ServicePortName, tableFamily util
 		svcRules.RuleID = svcRules.RuleID[:0]
 	}
 
-	return nil
-}
-
-func (p *proxy) deleteEndpointRules(ipTableFamily utilnftables.TableFamily, cn string, ruleID []uint64, svcPortName ServicePortName, key *epKey) error {
-	if err := nftables.DeleteEndpointRules(p.nfti, ipTableFamily, cn, ruleID); err != nil {
-		return err
-	}
-	// Deleting endpoint's chain
-	if err := nftables.DeleteChain(p.nfti, ipTableFamily, cn); err != nil {
-		klog.Errorf("failed to delete endpoint chain: %s with error: %+v", cn, err)
-		return err
-	}
-	// Check if it was last endpoint for a service port name
-	if len(p.endpointsMap[svcPortName]) == 0 {
-		klog.V(5).Infof("no more endpoints found for %s", svcPortName.String())
-		// No endpoints for svcPortName key is available, need to add svcPortName to No Endpoint Set
-		delete(p.endpointsMap, svcPortName)
-	}
-	return nil
-}
-
-func (p *proxy) addEndpoint(svcPortName ServicePortName, addr *v1.EndpointAddress, port *v1.EndpointPort) error {
-	p.mu.Lock()
-	defer p.mu.Unlock()
-	isLocal := addr.NodeName != nil && *addr.NodeName == p.hostname
-	ipFamily, ipTableFamily := getIPFamily(addr.IP)
-	baseEndpointInfo := newBaseEndpointInfo(ipFamily, port.Protocol, addr.IP, int(port.Port), isLocal, nil)
-	// Adding to endpoint base information, structures to carry nftables related info
-	baseEndpointInfo.epnft = &nftables.EPnft{
-		Interface: p.nfti,
-		Rule:      make(map[utilnftables.TableFamily]*nftables.EPRule),
-	}
-	cn := servicePortEndpointChainName(svcPortName.String(), string(port.Protocol), baseEndpointInfo.Endpoint)
-	// Initializing ip table family depending on endpoint's family ipv4 or ipv6
-	epRule := nftables.EPRule{
-		EpIndex: len(p.endpointsMap[svcPortName]),
-	}
-	epRule.Chain = cn
-	// RuleID nil is indicator that the nftables rule has not been yet programmed, once it is programed
-	// RuleID will be updated to real value.
-	epRule.RuleID = nil
-	// Check if corresponding ServicePort has Service Affinity set and copy parameters to endpoint rule struct
-	epRule.WithAffinity = false
-	if svc, ok := p.serviceMap[svcPortName]; ok {
-		epRule.WithAffinity = svc.(*serviceInfo).svcnft.WithAffinity
-		epRule.MaxAgeSeconds = svc.(*serviceInfo).svcnft.MaxAgeSeconds
-		epRule.ServiceID = svc.(*serviceInfo).svcnft.ServiceID
-	}
-	baseEndpointInfo.epnft.Rule[ipTableFamily] = &epRule
-	if err := p.addEndpointRules(&epRule, ipTableFamily, cn, svcPortName, &epKey{port.Protocol, addr.IP, port.Port}); err != nil {
-		klog.Errorf("failed to add endpoint rules for Service Port Name: %+v with error: %+v", svcPortName, err)
-		return err
-	}
-	p.endpointsMap[svcPortName] = append(p.endpointsMap[svcPortName], newEndpointInfo(baseEndpointInfo, port.Protocol))
-	if err := p.updateServiceChain(svcPortName, ipTableFamily); err != nil {
-		klog.Errorf("failed to update service %s chain with endpoint rule with error: %+v", svcPortName.String(), err)
-		return err
-	}
 	return nil
 }
